backend/requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# MQTT
paho-mqtt==1.6.1

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Data Validation
pydantic==2.6.1
pydantic-settings==2.1.0

# Configuration
PyYAML==6.0.1

# Database
aiosqlite==0.19.0

# Scheduling
APScheduler==3.10.4

# WebSocket
websockets==12.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.1.0
flake8>=7.0.0

# Deployment
pyinstaller>=6.0.0
